elab/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
cachetools==4.2.4
decorator==4.4.2
Django
django-appconf==1.0.5
django-compressor==3.1
django-crispy-forms==1.14.0
django-sass-processor==1.1
django-unicorn==0.42.1
importlib-metadata==4.11.1
libsass==0.21.0
Markdown==3.3.6
orjson==3.6.7
rcssmin==1.1.0
rjsmin==1.2.0
shortuuid==1.0.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
unicorn==1.0.3
zipp==3.7.0
